requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1
django-qr-code==4.1.0
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
segno==1.6.1
sqlparse==0.5.1
typing_extensions==4.12.2
